fix(codeql): numeric semantics — epsilon guards, cast-before-multiply, fraction preservation
cs/equality-on-floats (3 changed, 1 deferred):
- NWCreature.cs:4355,4375,4381: div-zero guards -> Math.Abs(x) < float.Epsilon.
  Provably bit-identical for binary32 (only ±0 has magnitude < Epsilon);
  boundary pinned in tests.
- NWCreature.cs:4357 (dLevel == 0f): DEFERRED for owner — result-check on
  the computed ratio (snap-to-1 = full XP); a practical epsilon would inflate
  XP. Gameplay-balance decision.

cs/loss-of-precision (5 changed):
- NWCreature.cs:1787, ProgressBar.cs:59, SoundEngine.cs:393: cast-before-
  multiply (int overflow before double cast). SoundEngine dx*dx overflowed
  for |d|>46340 (large coords -> NaN volume) — genuine bug, red/green proven.
- NWCreature.cs:514, BaseMainWindow.cs:353: fraction lost in int division —
  cast operands before dividing (33/100: 0% -> 33%; FPS 3/2000: 1 -> 1.5).
  Intended fixes per owner-approved golden tests.

Test: NumericRoundtripTests (9 tests, golden values computed via IEEE-754
emulation, pinned before edits). Red/green proven on the overflow case.
Registered in explicit <Compile> ItemGroup.

Verified: xbuild 0 errors; Tier B + C green.
Closes #79.

+++ project/Game/SoundEngine.cs
@@ -390,7 +390,7 @@ namespace NWR.Game
 
             int dx = player.X - sound.X;
             int dy = player.Y - sound.Y;
-            double dist = Math.Sqrt(dx * dx + dy * dy);
+            double dist = Math.Sqrt((double)dx * dx + (double)dy * dy);
             double scale = Math.Max(0.0, 1.0 - dist / MaxDistance);
             return (int)(volume * scale);
         }
